requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Naked==0.1.31
pycryptodomex==3.5.1
PyYAML==3.12
qiniu==7.2.0
requests==2.18.4
shellescape==3.4.1
SQLAlchemy==1.2.6
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
